requirements.txt
absl-py==1.2.0
astunparse==1.6.3
autoattack @ git+https://github.com/fra31/auto-attack.git@6482e4d6fbeeb51ae9585c41b16d50d14576aadc
cachetools==5.2.0
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
chardet==4.0.0
contourpy==1.0.5
cycler==0.11.0
flatbuffers==2.0.7
fonttools==4.37.3
gast==0.4.0
geotorch==0.3.0
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
idna==2.10
imageio==2.22.0
importlib-metadata==4.12.0
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
networkx==2.8.6
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.5
Pillow==9.2.0
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.0
pyparsing==3.0.9
pymoo
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
requests==2.25.1
requests-oauthlib==1.3.1
robustbench @ git+https://github.com/RobustBench/robustbench.git@eecfdef49ab30b62f2e6bcff3e3fe045b054557c
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
six==1.16.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tifffile==2022.8.12
torch==1.12.1
torchdiffeq==0.2.3
torchvision==0.13.1
tqdm==4.56.2
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
